add contactsController
