+++ routes/api/contacts.js
@@ -1,81 +1,34 @@
 const express = require('express')
 const router = express.Router()
+// const {
+//   listContacts,
+//   getContactById,
+//   removeContact,
+//   addContact,
+//   updateContact,
+// } = require('../../model/index')
+
 const {
-  listContacts,
-  getContactById,
-  removeContact,
-  addContact,
-  updateContact,
-} = require('../../model/index')
+  getContactList,
+  getContact,
+  postContact,
+  deleteContact,
+  patchContact
+} = require('../../src/controllers/contactsController')
 
 const {
   validationCreateContact,
   validationUpdateContact,
 } = require('../../src/middlewares/validation')
 
-router.get('/', async (req, res, next) => {
-  try {
-    const data = await listContacts()
-    return res.json({ data, status: 'success' })
-  } catch (err) {
-    console.log(err.message)
-  }
-  next()
-})
+router.get('/', getContactList)
 
-router.get('/:contactId', async (req, res, next) => {
-  const { contactId } = req.params
-  const id = Number(contactId)
-  try {
-    const data = await getContactById(id)
-    if (data) {
-      return res.status(200).json({ data, status: 'success' })
-    }
-    return res.status(404).json({ message: 'Not found' })
-  } catch (err) {
-    console.log(err.message)
-  }
-  next()
-})
+router.get('/:contactId', getContact)
 
-router.post('/', validationCreateContact, async (req, res, next) => {
-  try {
-    const data = await addContact(req.body)
-    return res.status(201).json({ data, status: 'success' })
-  } catch (err) {
-    console.log(err.message)
-  }
-  next()
-})
+router.post('/', validationCreateContact, postContact)
 
-router.delete('/:contactId', async (req, res, next) => {
-  const { contactId } = req.params
-  const id = Number(contactId)
-  try {
-    const data = await removeContact(id)
-    if (data) {
-      return res.status(200).json({ data, status: 'success' })
-    }
-    return res.status(404).json({ message: 'Not found' })
-  } catch (err) {
-    console.log(err.message)
-  }
-  next()
-})
+router.delete('/:contactId', deleteContact)
 
-router.patch('/:contactId', validationUpdateContact, async (req, res, next) => {
-  const { contactId } = req.params
-  const id = Number(contactId)
-  try {
-    const data = await updateContact(id, req.body)
-    if (data) {
-      return res.status(200).json({ data, status: 'success' })
-    }
-    return res.status(404).json({ message: 'Not found' })
-  } catch (err) {
-    console.log(err.message)
-  }
-  next()
-})
+router.patch('/:contactId', validationUpdateContact, patchContact)
 
 module.exports = router
